+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * Add `Size` properties to `Message`
 * `Nav` has been dreprecated in Bulma. Users should use `Navbar` from now.
 * Remove `FSharp.NET.Sdk` from the dependencies.
+* Implement `Navbar` components with documentation
 
 ### 0.1.0-alpha9