+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 * Remove `FSharp.NET.Sdk` from the dependencies.
 * Implement `Navbar` components with documentation
 * Implement `Pagination` components with documentation
+* Add explanation of how to use and and the architecture. (See Home page)
 
 ### 0.1.0-alpha9